requirements.txt
flake8==3.8.4
black==20.8b1
isort==5.6.4
pytest==6.1.2
mypy==0.790
fastapi==0.61.1
uvicorn==0.12.2
requests==2.24.0
PyYAML==5.3.1
python-multipart==0.0.5
python-jose==3.2.0
passlib==1.7.4
bcrypt==3.2.0
email_validator==1.1.2
emails==0.5.15
pytest-asyncio==0.14.0
pytest-cov==2.10.1
pandas==1.2.0
numpy==1.22.0
scikit-learn==0.24.1
sparse-dot-topn==0.2.9
sqlalchemy==1.3.22
psycopg2==2.8.6
pytest-env==0.6.2
xlrd==2.0.1
openpyxl==3.0.7
psutil==5.8.0
rq==1.8.0
redis==3.5.3
# fsspec==2021.11.1
s3fs==2021.8.0
awswrangler==2.12.0
# aiobotocore==2.0.1
behave==1.2.6
alembic==1.6.5
kubernetes==17.17.0
parameterized==0.8.1
tenacity==6.3.1
isodate==0.6.0
prometheus-client==0.7.1
